Move getMonacoContentWidth to the util file
